requirements.txt
flask
flask-sqlalchemy
requests
python-dotenv
flask-apscheduler